feat: Add type converter for String List and create FilterRule entity
This commit introduces a new `StringListConverter` to handle the conversion of lists of strings for storage in the Room database.
It also creates `FilterRuleEntity` to represent a filter rule in the local database and it utilizes the created `StringListConverter` to manage `List<String>` fields.
`SmsMessageEntity` is updated to use the `StringListConverter` type converter.